### 2018-03-12 a364db0
Remove useless line in init.py
