+++ ezyrb/__init__.py
@@ -4,32 +4,17 @@ __all__ = [
     'parametricspace', 'interpolation', 'rbf'
 ]
 
-try:
-    from . import filehandler
-    from . import matlabhandler
-    from . import interpolation
-    from . import vtkhandler
-    from . import stlhandler
-    from . import podinterpolation
-    from . import parametricspace
-    from . import online
-    from . import offline
-    from . import mapper
-    from . import points
-    from . import snapshots
-    from . import utilities
-    from .ndinterpolator import rbf
-except:
-    import filehandler
-    import matlabhandler
-    import interpolation
-    import vtkhandler
-    import stlhandler
-    import podinterpolation
-    import parametricspace
-    import online
-    import offline
-    import utilities
-    import mapper
-    import points
-    import snapshots
+from . import filehandler
+from . import matlabhandler
+from . import interpolation
+from . import vtkhandler
+from . import stlhandler
+from . import podinterpolation
+from . import parametricspace
+from . import online
+from . import offline
+from . import mapper
+from . import points
+from . import snapshots
+from . import utilities
+from .ndinterpolator import rbf
