+++ joss-paper/paper.bib
@@ -2,7 +2,7 @@
 %% http://bibdesk.sourceforge.net/
 
 
-%% Created for Marco Tezzele at 2018-03-07 17:26:55 +0100 
+%% Created for Marco Tezzele at 2018-03-15 12:02:05 +0100 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
@@ -21,7 +21,8 @@
 	Author = {Bui-Thanh, T and Damodaran, Murali and Willcox, Karen},
 	Booktitle = {21st AIAA Applied Aerodynamics Conference},
 	Date-Added = {2018-03-06 15:48:18 +0000},
-	Date-Modified = {2018-03-06 15:48:18 +0000},
+	Date-Modified = {2018-03-15 11:01:54 +0000},
+	Doi = {10.2514/6.2003-4213},
 	Pages = {4213},
 	Title = {Proper orthogonal decomposition extensions for parametric applications in compressible aerodynamics},
 	Year = {2003}}
@@ -37,7 +38,8 @@
 @book{hesthaven2016certified,
 	Author = {Hesthaven, Jan S and Rozza, Gianluigi and Stamm, Benjamin and others},
 	Date-Added = {2018-03-06 10:02:48 +0000},
-	Date-Modified = {2018-03-06 10:02:48 +0000},
+	Date-Modified = {2018-03-15 10:58:41 +0000},
+	Doi = {10.1007/978-3-319-22470-1},
 	Publisher = {Springer},
 	Title = {Certified reduced basis methods for parametrized partial differential equations},
 	Year = {2016}}
@@ -45,7 +47,8 @@
 @book{schilders2008model,
 	Author = {Schilders, Wilhelmus HA and Van der Vorst, Henk A and Rommes, Joost},
 	Date-Added = {2018-03-06 10:01:43 +0000},
-	Date-Modified = {2018-03-06 10:01:43 +0000},
+	Date-Modified = {2018-03-15 11:01:29 +0000},
+	Doi = {10.1007/978-3-540-78841-6},
 	Publisher = {Springer},
 	Title = {Model order reduction: theory, research aspects and applications},
 	Volume = {13},
@@ -55,7 +58,8 @@
 	Author = {Wojciak, Johannes and Theissen, Pascal and Adams, Nikolaus and Indinger, Thomas and Demuth, R},
 	Booktitle = {29th AIAA Applied Aerodynamics Conference},
 	Date-Added = {2018-03-06 09:20:43 +0000},
-	Date-Modified = {2018-03-06 09:20:43 +0000},
+	Date-Modified = {2018-03-15 10:59:25 +0000},
+	Doi = {10.2514/6.2011-3349},
 	Pages = {3349},
 	Title = {Investigation of Unsteady Vehicle Aerodynamics under Time-Dependent Flow Conditions},
 	Year = {2011}}
