### 2018-03-15 #614 @mtezzele
I've made a small fix to the bib file (missing doi).
