+++ docs/source/conf.py
@@ -43,6 +43,7 @@ extensions = [
     'sphinx.ext.doctest',
     'sphinx.ext.intersphinx',
     'sphinx.ext.todo',
+    'sphinx.ext.viewcode',
     'sphinx.ext.coverage',
     'sphinx.ext.imgmath',
     'sphinx.ext.ifconfig',
@@ -128,6 +129,13 @@ keep_warnings = False
 # If true, `todo` and `todoList` produce output, else they produce nothing.
 todo_include_todos = True
 
+# -- Options for viewcode extension ---------------------------------------
+
+# Follow alias objects that are imported from another module such as functions,
+# classes and attributes. As side effects, this option ... ???
+# If false, ... ???.
+# The default is True.
+viewcode_import = True
 
 # -- Options for HTML output ----------------------------------------------
 
