+++ _sources/contributing.rst.txt
@@ -17,7 +17,7 @@ Submitting a patch:
      separate branches in order to ensure that a pull request only includes the
      commits related to that bug or feature.
 
-  3. To ensure properly formatted code, please make sure to use a tab of 4
+  3. To ensure properly formatted code, please make sure to use 4
      spaces to indent the code. You should also run pylint over your code.
      It's not strictly necessary that your code be completely "lint-free",
      but this will help you find common style issues.
