+++ _sources/contact.rst.txt
@@ -3,5 +3,5 @@ Contact
 
 Feel free to contact the authors for any informations.
 
-- filippo.salmoiraghi@gmail.com
+- demo.nicola@gmail.com
 - marcotez@gmail.com
