### 2018-03-22 fcc8151
Add recent work in README
