### 2018-03-28 #614 @arfon
@ziotom78, @JaroslavHron, @dlagrava - would you be willing to review this submission for JOSS?
