### 2018-04-04 #614 @arfon
> @arfon I need to finish another review and I will be short of time during Easter. If the other guys do not take it or are too busy, I will take it.

@dlagrava ideally I'd like to find two reviewers for this submission as it's outside of my subject expertise and it seems like @jakevdp isn't going to be able to handle this one.
