### 2018-04-04 #661 @whedon
[REVIEW]: EZyRB: Easy Reduced Basis method
**Submitting author:** @mtezzele (<a href="http://orcid.org/0000-0001-9747-6328">Marco Tezzele</a>)
**Repository:** <a href="https://github.com/mathLab/EZyRB" target ="_blank">https://github.com/mathLab/EZyRB</a>
**Version:** v0.2.0
