### 2018-04-04 #614 @arfon
@dlagrava, @ziotom78 - please head over to #661 to carry out the actual JOSS review.
