+++ joss-paper/paper.bib
@@ -27,13 +27,14 @@
 	Title = {Proper orthogonal decomposition extensions for parametric applications in compressible aerodynamics},
 	Year = {2003}}
 
-@incollection{chinesta2016model,
-	Author = {Chinesta, Francisco and Huerta, Antonio and Rozza, Gianluigi and Willcox, Karen},
-	Date-Added = {2018-03-06 15:47:10 +0000},
-	Date-Modified = {2018-03-06 15:47:10 +0000},
-	Publisher = {Wiley},
-	Title = {Model Order Reduction: a survey},
-	Year = {2016}}
+@inbook{ChinestaHuertaRozzaWillcox2016,
+    chapter = {Model Order Reduction: a survey},
+    year = {2016},
+    publisher = {Wiley Encyclopedia of Computational Mechanics},
+    organization = {Wiley Encyclopedia of Computational Mechanics},
+    author = {Francisco Chinesta and Antonio Huerta and Gianluigi Rozza and Karen Willcox},
+    preprint = {https://urania.sissa.it/xmlui/bitstream/handle/1963/35194/ECM_MOR.pdf?sequence=1&isAllowed=y}
+    doi = {http://eu.wiley.com/WileyCDA/WileyTitle/productCd-1119003792.html}}
 
 @book{hesthaven2016certified,
 	Author = {Hesthaven, Jan S and Rozza, Gianluigi and Stamm, Benjamin and others},
