+++ joss-paper/paper.bib
@@ -33,7 +33,7 @@
     publisher = {Wiley Encyclopedia of Computational Mechanics},
     organization = {Wiley Encyclopedia of Computational Mechanics},
     author = {Francisco Chinesta and Antonio Huerta and Gianluigi Rozza and Karen Willcox},
-    preprint = {https://urania.sissa.it/xmlui/bitstream/handle/1963/35194/ECM_MOR.pdf?sequence=1&isAllowed=y}
+    preprint = {https://urania.sissa.it/xmlui/bitstream/handle/1963/35194/ECM_MOR.pdf?sequence=1&isAllowed=y},
     doi = {http://eu.wiley.com/WileyCDA/WileyTitle/productCd-1119003792.html}}
 
 @book{hesthaven2016certified,
