### 2018-04-06 #661 @mtezzele
@ziotom78 thanks for the suggestions. @ndem0 made all the fixes.
