### 2018-04-06 #661 @ziotom78
@mtezzele @ndem0 thank you for the prompt response! Everything seems to have been fixed, except the URL to Chinesta's book: it looks like it has a spurious `https://dx.doi.org/` at the beginning of it.

Might you please add the link to the arXiv entry for the paper by [Salmoiraghi et al](https://arxiv.org/abs/1803.04688)? Meanwhile, I'll play with EZyRB, I just installed it on my computer.
