### 2018-04-06 #661 @ndem0
@ziotom78 I fixed both the wrong doi and arXiv link missing, now the paper should be ok!
