### 2018-04-07 #661 @dlagrava
@mtezzele a couple of typos on the paper.md:
- since its structure can interact with several simulation software **by** simply providing the output file of the simulations
- As an **example**, ...

Thanks.
