### 2018-04-09 mathLab/EZyRB#68 @ziotom78
Clearly state which Python version is supported
The README does not tell whether Python 2 or 3 is supported. EZyRB currently can be installed either using Python 2 or 3, but tests fail if Python 3.6 is used. The file `setup.py` states that Python 2.7 is the only version supported.
