### 2018-04-09 #661 @ziotom78
I created [ticket #68](https://github.com/mathLab/EZyRB/issues/68) on EZyRB's website, as there is no clear indication of the supported Python version. (I discovered that EZyRB does not support Python 3.6 by trial and error, although I later realized that `setup.py` indicates that in the metadata.)
