### 2018-04-09 mathLab/EZyRB#68 @ndem0
Add information about the supported python version in #69.
