### 2018-04-09 #661 @ndem0
@ziotom78 Thanks for the segnalation, I added in the README file the supported Python version.
