### 2018-04-10 #661 @arfon
@dlagrava, @ziotom78 - can you both confirm we're good to accept here now?
